requirements.txt
python-dotenv==1.0.0
requests==2.32.4
streamlit==1.37.0
deep-translator==1.11.4
